backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Admin==1.6.1
WTForms==3.1.2
python-dotenv==1.0.0
werkzeug==3.0.1

